requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
tzdata==2021.5
